requirements.txt
textual
numpy
scipy
hmmlearn
yfinance
fredapi
duckdb
requests
python-dotenv
pytest
